DjangoServer/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
dj_database_url
gunicorn
django-heroku
whitenoise
djangorestframework==3.12.1   
django-rest-authtoken==2.1.3
django-cors-headers==3.5.0  
pandas==1.0.1
requests==2.22.0 
boto3==1.15.18